refactor(monorepo): create pre package
