+++ README.md
@@ -1,1 +1,37 @@
-# nucypher-ts-mono
+# nucypher-ts
+
+A TypeScript library for interacting with access control functionality in the browser.
+
+Full documentation can be found [here](https://docs.threshold.network/app-development/threshold-access-control-tac).
+
+> **Warning**
+>
+> `nucypher-ts` is under [active development](https://github.com/nucypher/nucypher-ts/pulls):
+>
+> - SDK does not support policy revocation.
+> - We expect breaking changes.
+
+## Installation
+
+```
+yarn add @nucypher/nucypher-ts
+```
+
+## Tutorial
+
+To learn more, follow the tutorial at Threshold Network's [docs](https://docs.threshold.network/app-development/threshold-access-control-tac/get-started-with-tac).
+
+## Examples
+
+See [`nucypher-ts/examples`](https://github.com/nucypher/nucypher-ts/tree/main/examples) to find out how to integrate `nucypher-ts` into your favorite web framework.
+
+We also provide two code samples of TAC applications:
+
+- [nucypher/tdec-sandbox](https://github.com/nucypher/tdec-sandbox)
+- [nucypher/tdec-nft-example](https://github.com/nucypher/tdec-nft-example)
+
+These examples showcase integration with React-based web application and an end-to-end flow of creating conditioned encryption, and encrypting & decrypting data.
+
+# Contributing
+
+If you would like to contribute to the development of `nucypher-ts`, please see our [Contributing Guide](CONTRIBUTING.md). You can also join our [Discord](http://discord.gg/threshold) and say hello!
