+++ packages/pre/package.json
@@ -1,0 +1,63 @@
+{
+  "name": "@nucypher/pre",
+  "version": "0.1.0",
+  "keywords": [
+    "pre",
+    "threshold",
+    "access control"
+  ],
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/nucypher/nucypher-ts"
+  },
+  "license": "GPL-3.0-only",
+  "author": "Piotr Roslaniec <p.roslaniec@gmail.com>",
+  "main": "./dist/index.js",
+  "module": "./dist/index.js",
+  "types": "./dist/index.d.ts",
+  "files": [
+    "dist/**/*"
+  ],
+  "scripts": {
+    "build": "tsc --build ./tsconfig.build.json --verbose",
+    "docs": "typedoc --json ../../docs-json/nucypher-ts.json --validation.invalidLink false",
+    "exports:lint": "ts-unused-exports tsconfig.json --ignoreFiles src/index.ts",
+    "postinstall": "yarn typechain",
+    "package-check": "package-check",
+    "prepare-release": "run-s reset-hard test doc:html version doc:publish",
+    "prepare-release:alpha": "run-s reset-hard test version:alpha ",
+    "test": "vitest run",
+    "version": "standard-version",
+    "version:alpha": "standard-version --prerelease alpha"
+  },
+  "dependencies": {
+    "@ethersproject/abi": "^5.7.0",
+    "@ethersproject/abstract-signer": "^5.7.0",
+    "@ethersproject/providers": "^5.7.2",
+    "@nucypher/nucypher-core": "^0.12.0",
+    "@nucypher/shared": "workspace:*",
+    "axios": "^1.5.0",
+    "deep-equal": "^2.2.1",
+    "ethers": "^5.7.2",
+    "qs": "^6.10.1",
+    "semver": "^7.5.2",
+    "vitest": "^0.34.4",
+    "zod": "^3.22.1"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.22.17",
+    "@babel/preset-env": "^7.22.15",
+    "@babel/preset-typescript": "^7.22.15",
+    "@typechain/ethers-v5": "^11.1.1",
+    "@types/deep-equal": "^1.0.1",
+    "@types/qs": "^6.9.7",
+    "@types/semver": "^7.5.0",
+    "cz-conventional-changelog": "^3.0.1",
+    "standard-version": "^9.0.0",
+    "typechain": "^8.3.1"
+  },
+  "engines": {
+    "node": ">=16",
+    "pnpm": ">=8.0.0"
+  }
+}
