+++ packages/pre/test/pre.test.ts
@@ -1,0 +1,6 @@
+import {test, expect} from 'vitest';
+
+
+test('pre', () => {
+  expect('pre').toBe('pre');
+});
