+++ packages/pre/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "../../tsconfig.json",
+  "include": ["src", "test"],
+  "compilerOptions": {
+    "esModuleInterop": true,
+    "skipLibCheck": true,
+  }
+}
