--- packages/taco/.czrc.json
@@ -1,3 +1,0 @@
-{
-  "path": "cz-conventional-changelog"
-}
